==> trajectory_plot_1/1_data_preprocess.ipynb <==
In this script, there are two main parts: detect the log type errors and clean up them.

For the detection part, there are 3 functions: 
1 The function can detect particant and scenario whose number of taskAccepted is not equal to 4. 
2 The function can detect particant and scenario whose number of checkedIn is not equal to 4. 
3 The function can detect particant and scenario whose number of navigation is not equal to 4. 

import plotmethods as pl
combined_dataframe =  pl.read_csv_files('rawdata')
print (combined_dataframe)

pl.filter_duplicated_taskAccepted(combined_dataframe)

#For participant31 scneario3, there are only 3 taskAccepted, no error.

pl.filter_duplicated_checkedIn(combined_dataframe)

pl.filter_navigation_groups(combined_dataframe)

For the clean up part. The wrong types of log will be repaired.


pl.replace_fourth_taskAccepted(combined_dataframe, 27, 1)
pl.replace_fourth_taskAccepted(combined_dataframe, 27, 3)
pl.replace_second_taskAccepted(combined_dataframe, 38, 3)
pl.replace_fifth_taskAccepted(combined_dataframe, 44, 1)
pl.replace_fourth_taskAccepted(combined_dataframe, 45, 3)
pl.replace_fifth_taskAccepted(combined_dataframe, 62, 2)
pl.replace_third_taskAccepted(combined_dataframe, 32, 1)


pl.replace_fourth_taskAccepted(combined_dataframe, 17, 3)
pl.replace_fourth_taskAccepted(combined_dataframe, 17, 3)

#check it For participant31 scneario3, there are only 3 taskAccepted, no error.
pl.filter_duplicated_taskAccepted(combined_dataframe)


pl.replace_fifth_checkedIn(combined_dataframe, 16, 3)
pl.replace_fifth_checkedIn(combined_dataframe, 19, 1)
pl.replace_fifth_checkedIn(combined_dataframe, 40, 1)
pl.replace_fifth_checkedIn(combined_dataframe, 53, 4)
pl.replace_fifth_checkedIn(combined_dataframe, 53, 4)
pl.replace_fifth_checkedIn(combined_dataframe, 53, 4)
pl.replace_fifth_checkedIn(combined_dataframe, 53, 4)
pl.replace_fifth_checkedIn(combined_dataframe, 59, 2)
pl.replace_third_checkedIn(combined_dataframe, 31, 3)


##check it For participant31 scneario3, there are only 3 checkedIn, no error.
pl.filter_duplicated_checkedIn(combined_dataframe)


pl.delete_map_interaction(17, 3, combined_dataframe)
pl.delete_map_interaction(17, 3, combined_dataframe)
pl.delete_map_interaction(17, 3, combined_dataframe)
pl.delete_map_interaction(27, 1, combined_dataframe)
pl.delete_map_interaction(27, 1, combined_dataframe)
pl.delete_map_interaction(27, 3, combined_dataframe)
pl.delete_map_interaction(31, 3, combined_dataframe)
pl.delete_map_interaction(38, 3, combined_dataframe)
pl.delete_map_interaction(44, 1, combined_dataframe)
pl.delete_map_interaction(45, 3, combined_dataframe)
pl.delete_map_interaction(59, 3, combined_dataframe)
pl.delete_map_interaction(59, 4, combined_dataframe)
pl.delete_map_interaction(62, 2, combined_dataframe)
pl.delete_map_interaction(62, 2, combined_dataframe)
pl.delete_map_interaction_second(31, 1, combined_dataframe)

  
pl.remove_duplicate_navigation_targets(combined_dataframe, 17, 3)
pl.remove_duplicate_navigation_targets(combined_dataframe, 32, 1)
pl.remove_duplicate_navigation_targets(combined_dataframe, 37, 1)
pl.remove_duplicate_navigation_targets(combined_dataframe, 59, 4)



updated_dataframe = pl.mark_start_end_state(combined_dataframe)


updated_dataframe = pl.mark_navigation_phase(updated_dataframe)

updated_dataframe = pl.mark_planning_phase(updated_dataframe)

updated_dataframe.to_csv('updated_dataframe.csv', index=False, encoding='utf-8')

Next, try to give a tag of routename to each row

pl.count_checkedIn_types(updated_dataframe)

updated_dataframe = pl.add_routename_column(updated_dataframe)

print(updated_dataframe.iloc[218218])

# Delete the row at position 218218
updated_dataframe.drop(index=218218, inplace=True)

updated_dataframe.to_csv('dataframe_allroutename.csv', index=False, encoding='utf-8')
