==> trajectory_plot_1/6_frequncy.ipynb <==
import pandas as pd

# Assuming your DataFrame is named df
df = pd.read_csv('df_all.csv')
dff = pd.read_csv('result_phase_analysis.csv')
# Filter the DataFrame
filtered_df = df[(df['phase'] == 'navigation') & (df['mapInteractions'] != '')]

# Group by scenario and participant, and count occurrences
result = filtered_df.groupby(['scenario', 'participant']).size().reset_index(name='interaction_count')

print(result)

df_checktimes = pd.merge(dff, result, on = ['participant', 'scenario'])
df_checktimes['frequency_distance'] = round(df_checktimes['interaction_count'] / (df_checktimes['planning_distance'] + df_checktimes['navigating_distance']),2)
df_checktimes['frequency_time'] = round(df_checktimes['interaction_count'] / (df_checktimes['planning_time'] + df_checktimes['navigation_time']),2)


df_checktimes.to_csv('result_checktimes.csv')
